separate api constants to config file
