+++ src/api/index.ts
@@ -2,4 +2,5 @@ export * from "./endpoints/employees.api";
 export * from "./endpoints/departments.api";
 export * from "./endpoints/roles.api";
 export * from "./endpoints/skills.api";
-export { default as API } from "./api";
+export * from "./services/api.config";
+export { default as API } from "./services/api";
