+++ src/api/services/api.config.ts
@@ -1,0 +1,4 @@
+export const apiConfig = {
+  baseURL: "http://3.145.178.76:4000/",
+  timeout: 120000,
+};
